+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # NEXT RELEASE
+* Create config.locale.messages for potentially user facing string configuration
+* Reset trendingData when reset() is called
 
 ### v0.74.2
 * Rename `TagTootsType` enum to `TagTootsCategory`